backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djoser>=2.2.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.2
psycopg2-binary>=2.9.9
python-decouple>=3.8
Pillow>=10.1.0
